Alterando tipo de valor

+++ src/main/java/com/example/Fibonacci.java
@@ -5,7 +5,7 @@ public class Fibonacci{
     public Fibonacci(){
 
     }
-    public static long fibo(int n) {
+    public static double fibo(int n) {
         if (n >= 0) {
             return Math.pow(2,n);
         } else {
